+++ modules/default/weather/weatherprovider.js
@@ -111,23 +111,17 @@ const WeatherProvider = Class.extend({
 		this.delegate.updateAvailable(this);
 	},
 
-	getCorsUrl: function () {
-		if (this.config.mockData || typeof this.config.useCorsProxy === "undefined" || !this.config.useCorsProxy) {
-			return "";
-		} else {
-			return location.protocol + "//" + location.host + "/cors?url=";
-		}
-	},
-
 	/**
 	 * A convenience function to make requests.
 	 *
 	 * @param {string} url the url to fetch from
 	 * @param {string} type what contenttype to expect in the response, can be "json" or "xml"
+	 * @param {Array.<{name: string, value:string}>} requestHeaders the HTTP headers to send
+	 * @param {Array.<string>} expectedResponseHeaders the expected HTTP headers to recieve
 	 * @returns {Promise} resolved when the fetch is done
 	 */
-	fetchData: async function (url, type = "json") {
-		url = this.getCorsUrl() + url;
+	fetchData: async function (url, type = "json", requestHeaders = undefined, expectedResponseHeaders = undefined) {
+		url = this.getCorsUrl(url, requestHeaders, expectedResponseHeaders);
 		const mockData = this.config.mockData;
 		if (mockData) {
 			const data = mockData.substring(1, mockData.length - 1);
@@ -135,12 +129,114 @@ const WeatherProvider = Class.extend({
 		} else {
 			const response = await fetch(url);
 			const data = await response.text();
+
 			if (type === "xml") {
 				return new DOMParser().parseFromString(data, "text/html");
 			} else {
-				return JSON.parse(data);
+				if (!data || !data.length > 0) return undefined;
+
+				const dataResponse = JSON.parse(data);
+				if (!dataResponse.headers) {
+					dataResponse.headers = this.getHeadersFromResponse(expectedResponseHeaders, response);
+				}
+				return dataResponse;
 			}
 		}
+	},
+
+	/**
+	 * Gets a URL that will be used when calling the CORS-method on the server.
+	 *
+	 * @param {string} url the url to fetch from
+	 * @param {Array.<{name: string, value:string}>} requestHeaders the HTTP headers to send
+	 * @param {Array.<string>} expectedResponseHeaders the expected HTTP headers to recieve
+	 * @returns {string} to be used as URL when calling CORS-method on server.
+	 */
+	getCorsUrl: function (url, requestHeaders, expectedResponseHeaders) {
+		if (this.config.mockData || typeof this.config.useCorsProxy === "undefined" || !this.config.useCorsProxy) {
+			return "";
+		} else if (!url || url.length < 1) {
+			throw new Error(`Invalid URL: ${url}`);
+		} else {
+			let corsUrl = `${location.protocol}//${location.host}/cors?`;
+
+			const requestHeaderString = this.getRequestHeaderString(requestHeaders);
+			if (requestHeaderString) corsUrl = `${corsUrl}sendheaders=${requestHeaderString}`;
+
+			const expectedResponseHeadersString = this.getExpectedResponseHeadersString(expectedResponseHeaders);
+			if (requestHeaderString && expectedResponseHeadersString) {
+				corsUrl = `${corsUrl}&expectedheaders=${expectedResponseHeadersString}`;
+			} else if (expectedResponseHeadersString) {
+				corsUrl = `${corsUrl}expectedheaders=${expectedResponseHeadersString}`;
+			}
+
+			if (requestHeaderString || expectedResponseHeadersString) {
+				return `${corsUrl}&url=${url}`;
+			}
+			return `${corsUrl}url=${url}`;
+		}
+	},
+
+	/**
+	 * Gets the part of the CORS URL that represents the HTTP headers to send.
+	 *
+	 * @param {Array.<{name: string, value:string}>} requestHeaders the HTTP headers to send
+	 * @returns {string} to be used as request-headers component in CORS URL.
+	 */
+	getRequestHeaderString: function (requestHeaders) {
+		let requestHeaderString = "";
+		if (requestHeaders) {
+			for (const header of requestHeaders) {
+				if (requestHeaderString.length === 0) {
+					requestHeaderString = `${header.name}:${encodeURIComponent(header.value)}`;
+				} else {
+					requestHeaderString = `${requestHeaderString},${header.name}:${encodeURIComponent(header.value)}`;
+				}
+			}
+			return requestHeaderString;
+		}
+		return undefined;
+	},
+
+	/**
+	 * Gets the part of the CORS URL that represents the expected HTTP headers to recieve.
+	 *
+	 * @param {Array.<string>} expectedResponseHeaders the expected HTTP headers to recieve
+	 * @returns {string} to be used as the expected HTTP-headers component in CORS URL.
+	 */
+	getExpectedResponseHeadersString: function (expectedResponseHeaders) {
+		let expectedResponseHeadersString = "";
+		if (expectedResponseHeaders) {
+			for (const header of expectedResponseHeaders) {
+				if (expectedResponseHeadersString.length === 0) {
+					expectedResponseHeadersString = `${header}`;
+				} else {
+					expectedResponseHeadersString = `${expectedResponseHeadersString},${header}`;
+				}
+			}
+			return expectedResponseHeaders;
+		}
+		return undefined;
+	},
+
+	/**
+	 * Gets the values for the expected headers from the response.
+	 *
+	 * @param {Array.<string>} expectedResponseHeaders the expected HTTP headers to recieve
+	 * @param {Response} response the HTTP response
+	 * @returns {string} to be used as the expected HTTP-headers component in CORS URL.
+	 */
+	getHeadersFromResponse(expectedResponseHeaders, response) {
+		const responseHeaders = [];
+
+		if (expectedResponseHeaders) {
+			for (const header of expectedResponseHeaders) {
+				const headerValue = response.headers.get(header);
+				responseHeaders.push({ name: header, value: headerValue });
+			}
+		}
+
+		return responseHeaders;
 	}
 });